@@ -21,39 +21,38 @@
     "test"
   ],
   "files": [
-    "cli.js",
-    "index.js"
+    "cli.mjs",
+    "index.mjs"
   ],
-  "bin": "cli.js",
-  "main": "index.js",
+  "bin": "cli.mjs",
+  "main": "index.mjs",
   "engines": {
     "node": "^12.20 || >= 14.13"
   },
   "dependencies": {
-    "fs-extra": "^8.1.0"
+    "fs-extra": "^10.0.0"
   },
   "devDependencies": {
-    "coverage-node": "^2.0.2",
-    "disposable-directory": "^1.0.0",
-    "eslint": "^6.8.0",
-    "eslint-config-env": "^13.0.3",
-    "eslint-config-prettier": "^6.9.0",
-    "eslint-plugin-import": "^2.20.0",
-    "eslint-plugin-jsdoc": "^21.0.0",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-prettier": "^3.1.2",
-    "jsdoc-md": "^5.0.1",
-    "prettier": "^1.19.1",
-    "test-director": "^3.0.1"
+    "coverage-node": "^5.0.1",
+    "disposable-directory": "^4.0.0",
+    "eslint": "^7.27.0",
+    "eslint-config-env": "^20.0.1",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.23.3",
+    "eslint-plugin-jsdoc": "^35.0.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^3.4.0",
+    "jsdoc-md": "^11.0.0",
+    "prettier": "^2.3.0",
+    "test-director": "^6.0.0"
   },
   "scripts": {
-    "prepare": "npm run prepare:jsdoc && npm run prepare:prettier",
-    "prepare:jsdoc": "jsdoc-md",
-    "prepare:prettier": "prettier readme.md --write",
-    "test": "npm run test:eslint && npm run test:prettier && npm run test:api",
+    "jsdoc": "jsdoc-md",
+    "test": "npm run test:eslint && npm run test:prettier && npm run test:jsdoc && npm run test:api",
     "test:eslint": "eslint .",
-    "test:prettier": "prettier '**/*.{json,yml,md}' -l",
-    "test:api": "coverage-node test",
+    "test:prettier": "prettier -c .",
+    "test:jsdoc": "jsdoc-md -c",
+    "test:api": "coverage-node test.mjs",
     "prepublishOnly": "npm test"
   }
 }
