@@ -1,9 +1,8 @@
-'use strict';
-
-const { execFile } = require('child_process');
-const { resolve } = require('path');
-const { promisify } = require('util');
-const { remove } = require('fs-extra');
+import { execFile } from 'child_process';
+import fs from 'fs';
+import { resolve } from 'path';
+import { promisify } from 'util';
+import { remove } from 'fs-extra';
 
 const execFilePromise = promisify(execFile);
 
@@ -14,10 +13,12 @@ const execFilePromise = promisify(execFile);
  * @param {string} pathPackageFrom Absolute or CWD relative filesystem path to the package to install from.
  * @param {string} [pathPackageTo] Absolute or CWD relative filesystem path to the package to install to, defaulting to `process.cwd()`.
  * @returns {Promise<void>} Resolves once installation is complete.
+ * @example <caption>How to `import`.</caption>
+ * ```js
+ * import installFrom from 'install-from';
+ * ```
  * @example <caption>Install a package into another.</caption>
  * ```js
- * const { installFrom } = require('install-from');
- *
  * installFrom(
  *   './packages/package-to-install-from',
  *   './packages/package-to-install-to'
@@ -30,12 +31,14 @@ const execFilePromise = promisify(execFile);
  *   });
  * ```
  */
-exports.installFrom = async function installFrom(
+export default async function installFrom(
   pathPackageFrom,
   pathPackageTo = process.cwd()
 ) {
   // Determine the package pack path.
-  const { name, version } = require(resolve(pathPackageFrom, 'package.json'));
+  const { name, version } = JSON.parse(
+    await fs.promises.readFile(resolve(pathPackageFrom, 'package.json'), 'utf8')
+  );
   const pathPack = resolve(
     pathPackageFrom,
 
@@ -48,9 +51,9 @@ exports.installFrom = async function installFrom(
 
   // Install the packed package.
   await execFilePromise('npm', ['install', pathPack, '--no-save'], {
-    cwd: pathPackageTo
+    cwd: pathPackageTo,
   });
 
   // Delete the pack.
   await remove(pathPack);
-};
+}
